+++ src/helpers.js
@@ -1,3 +1,11 @@
+/*
+	A helper should honor the current rule properties, including:
+	- continue()
+	- next()
+
+	Best practices:
+	- name the handlers according to the helper name (useful in debug mode)
+**/
 var Atok = require('atok')
 var isArray = require('util').isArray
 
@@ -5,8 +13,8 @@ module.exports._helper_setArguments = function (defaults, args, type) {
 	var atok = this, n = args.length
 	var res = [].concat(defaults)
 
-	var	defaultHandler = atok.handler || function (token, idx) {
-		atok.emit('data', token, idx, type)
+	var	defaultHandler = atok.handler || function helperDefaultHandler (token) {
+		atok.emit_data(token, arguments.length > 1 ? arguments[1] : -1, type)
 	}
 
 	if (n === 0)
@@ -28,10 +36,11 @@ module.exports._helper_word = function (delimiters, handler, wordStart) {
 	var atok = this
 	var current = atok.helpersCache._helper_word || ''
 
-	function acc (data) {
+	function _helper_wordAcc (data) {
 		current += data
 	}
-	function done () {
+	function _helper_wordDone () {
+		atok.seek(-1) // Get back by one as it was artificially added
 		handler(current, -1, null)
 		atok.helpersCache._helper_word = current = ''
 	}
@@ -54,24 +63,26 @@ module.exports._helper_word = function (delimiters, handler, wordStart) {
 				}
 			, handler
 			)
+			.loadProps('_helper_word')
 	else
 		atok
+			.next()
 			// while(character matches a word letter)
 			.continue(-1)
-				.addRule(wordStart, 0, function (token) {
+				.addRule(wordStart, 0, function _helper_wordWait (token) {
 					// Character matches but end of buffer reached
-					acc(token)
-					if (atok.ending) done()
+					_helper_wordAcc(token)
+					if (atok.ending) _helper_wordDone()
 					// Since continue() is used, the rule index is preserved
 					else atok.helpersCache.word = current
 				})
 			.continue(-2)
-				.addRule(wordStart, acc)
+				.addRule(wordStart, _helper_wordAcc)
 			.loadProps('_helper_word')
 			.addRule(function () {
-				return current.length > 0 ? 0 : -1
-			}, done)
-	
+				return current.length > 0 ? 1 : -1
+			}, _helper_wordDone)
+
 	return atok
 }