+++ src/parser.js
@@ -2,15 +2,17 @@ var fs = require('fs')
 var path = require('path')
 var Stream = require('stream').Stream
 var vm = require('vm')
+var isArray = require('util').isArray
 
 var EV = require('ev')
 var inherits = require('inherits')
 var requireLike = require('require-like')
-var Atok = require('atok')
+// var Atok = require('atok')
+var Atok = require('../../node-atok')
 var Helpers = require('./helpers')
 var Tracker = require('./tracker')
 
-function noOp () {}
+function noop () {}
 
 // Copy properties from one object to another - not replacing if required
 function merge (a, b, soft) {
@@ -32,16 +34,19 @@ function inspect (s) {
 merge(Atok.prototype, Helpers)
 exports.Helpers = Helpers
 
+// Export version
+exports.version = require('../package.json').version
+
 /**
- * createParser(file, parserOptions, parserEvents, atokOptions)
- * - file (String): file to read the parser from(.js extension is optional)
+ * createParserFromContent(data, parserOptions, parserEvents, atokOptions)
+ * - data (String | Array): parser javascript code
  * - parserOptions (String): list of the parser named events with their arguments count
  * - parserEvents (Object): events emitted by the parser with
  * - atokOptions (Object): tokenizer options
  *
  * Return a parser class (Function) based on the input file.
  * The following variables are made available to the parser js:
- * - atok {Object}: atok tokenizer instanciated with provided options. Also 
+ * - atok (Object): atok tokenizer instanciated with provided options. Also 
  *   set as this.atok *DO NOT DELETE*
  * - self (Object): reference to this
  * Predefined methods:
@@ -52,15 +57,26 @@ exports.Helpers = Helpers
  * - track
  * Events automatically forwarded from tokenizer to parser:
  * - drain
- * - debug
 **/
-exports.createParser = function (file, parserOptions, parserEvents, atokOptions) {
-  var filename = file.slice(-3) === '.js' ? file : file + '.js'
-  filename = path.resolve( path.dirname(module.parent.filename), filename )
-
+exports.createParserFromContent = function (data, parserOptions, parserEvents, atokOptions, filename) {
+  filename = filename || ''
   // Merge the supplied events with Atok's (overwrite)
   var _parserEvents = merge(parserEvents || {}, Atok.events)
 
+  // Set an event arguments list
+  function forwardEvent (event) {
+    var n = Atok.events[event]
+      , args = []
+
+    while (n > 0) args.push('a' + n--)
+    args = args.join()
+
+    return 'atok.on("'
+      + event
+      + '", function (' + args + ') { self.emit_drain(' + args + ') })'
+  }
+
+  // Define the parser constructor
   var content = []
   content.push(
     'function Parser (' + (parserOptions || '') + ') {'
@@ -71,12 +87,12 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
   , 'this.writable = true'
   , 'var self = this'
   , 'var atok = new Atok(' + (atokOptions ? JSON.stringify(atokOptions): '') + ')'
-  , 'atok.on("drain", this.emit_drain)'
-  , 'atok.on("debug", this.emit_debug)'
+  , forwardEvent('drain')
+  , forwardEvent('debug')
   , 'atok.helpersCache = {}'
   , 'this.atok = atok'
   , 'this.atokTracker = new Tracker(atok)'
-  , fs.readFileSync(filename).toString()
+  , isArray(data) ? data.join(';') : data
   , '}'
   , 'inherits(Parser, EV, Stream.prototype)'
   , 'exports = Parser'
@@ -123,9 +139,10 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
   Parser.prototype.end = function (data) {
     return this.atok.end(data)
   }
-  Parser.prototype.destroy = noOp
+  Parser.prototype.destroy = noop
   // Track current line and column
   // type: {String | Boolean} platform type (used to identify newline characters)
+  // Note. newline characters are automatically set by looking at process.platform()
   Parser.prototype.track = function (type) {
     if (type) { // Turn tracking ON
       this.atokTracker.start(type)
@@ -147,8 +164,7 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
 
     var tracker = this.atokTracker
 
-    var msg = ''
-    msg += 'rule: ' + atok.getRuleSet()
+    var msg = 'rule: ' + atok.getRuleSet()
 
     if (tracker.running) {
       var margin = 4
@@ -169,8 +185,10 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
       // Position of the cursor below the displayed data
       var cursor = pos - left + margin + leftInspected.length - leftExtract.length
 
-      msg += ', line: ' + line
-      msg += ', column: ' + column
+      msg += ', line: '
+      msg += line
+      msg += ', column: '
+      msg += column
       msg += '\n'
       // Data sample around the current buffer position
       msg += Array(margin + 1).join(' ')
@@ -182,7 +200,8 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
       // Add the cursor
       msg += Array((Math.min(size, right - pos) || 1) + 1).join('^')
     } else {
-      msg += ', token: ' + inspect( data )
+      msg += ', token: '
+      msg += inspect( data )
     }
     msg += '\n'
 
@@ -196,4 +215,27 @@ exports.createParser = function (file, parserOptions, parserEvents, atokOptions)
   }
 
   return Parser
+}
+
+/**
+ * createParser(file, parserOptions, parserEvents, atokOptions)
+ * - file (String): file to read the parser from(.js extension is optional but enforced)
+ * - parserOptions (String): list of the parser named events with their arguments count
+ * - parserEvents (Object): events emitted by the parser with
+ * - atokOptions (Object): tokenizer options
+**/
+exports.createParser = function (file, parserOptions, parserEvents, atokOptions) {
+  // Append .js extension if not set
+  var filename = file.slice(-3) === '.js' ? file : file + '.js'
+  // Figure out the file full path
+  filename = path.resolve( path.dirname(module.parent.filename), filename )
+
+  return exports
+    .createParserFromContent(
+      fs.readFileSync(filename).toString()
+    , parserOptions
+    , parserEvents
+    , atokOptions
+    , filename
+    )
 }